// packages/uns-core/src/graphql/schema.d.ts
export declare enum UnsAttributeType {
    Api = "Api",
    Data = "Data",
    Event = "Event",
    Table = "Table"
}